requirements.txt
Django==2.0
psycopg2==2.7.5
PyJWT==1.6.4
gunicorn==19.9.0
whitenoise==4.1.3
firebase-admin==2.13.0
ipython==6.5.0
google-cloud-storage==1.19.1
# django-heroku==0.3.1
